Revert "Lesson 2.8"
