+++ src/main/java/com/Horns/Hooves/Employee/Employee.java
@@ -1,0 +1,44 @@
+package com.Horns.Hooves.Employee;
+
+import java.util.Objects;
+
+public class Employee {
+    private String name;
+    private String surname;
+
+
+    public Employee(String name,String surname) {
+        this.name = name;
+        this.surname = surname;
+
+
+    }
+
+    public String getName() {
+        return name;
+    }
+    public String getSurname() {return surname;}
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Employee employee = (Employee) o;
+        return name.equals(employee.name) && surname.equals(employee.surname);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name, surname);
+    }
+
+    @Override
+    public String toString() {
+        return "Employee{" +
+                "name='" + name + '\'' +
+                ", surname='" + surname + '\'' +
+                '}';
+    }
+}
+
+
