+++ src/main/java/com/Horns/Hooves/Employee/EmployeeAlreadyAddedException.java
@@ -1,0 +1,7 @@
+package com.Horns.Hooves.Employee;
+
+public class EmployeeAlreadyAddedException extends RuntimeException{
+    public EmployeeAlreadyAddedException(String messega){
+        super(messega);
+    }
+}
