+++ src/main/java/com/Horns/Hooves/Employee/EmployeeNotFoundException.java
@@ -1,0 +1,7 @@
+package com.Horns.Hooves.Employee;
+
+public class EmployeeNotFoundException extends RuntimeException{
+    public EmployeeNotFoundException(String messega){
+        super(messega);
+    }
+}
