+++ src/main/java/com/Horns/Hooves/Employee/EmployeeServiceInterf.java
@@ -1,0 +1,7 @@
+package com.Horns.Hooves.Employee;
+
+public interface EmployeeServiceInterf {
+    // Добавление сотрудника
+    void addEmployee(Employee employee) throws EmployeeStorageIsFullException,EmployeeAlreadyAddedException;
+    void findEmployee(Employee employee) throws EmployeeNotFoundException;
+}
