+++ src/main/java/com/Horns/Hooves/Employee/EmployeeStorageIsFullException.java
@@ -1,0 +1,7 @@
+package com.Horns.Hooves.Employee;
+
+public class EmployeeStorageIsFullException extends RuntimeException {
+    public EmployeeStorageIsFullException(String messega) {
+        super(messega);
+    }
+}
