--- src/main/java/com/Horns/Hooves/Employee/exception/EmployeeAlreadyAddedException.java
@@ -1,25 +1,0 @@
-package com.Horns.Hooves.Employee.exception;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
-public class EmployeeAlreadyAddedException extends RuntimeException{
-    public EmployeeAlreadyAddedException() {
-    }
-
-    public EmployeeAlreadyAddedException(String message) {
-        super(message);
-    }
-
-    public EmployeeAlreadyAddedException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public EmployeeAlreadyAddedException(Throwable cause) {
-        super(cause);
-    }
-
-    public EmployeeAlreadyAddedException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
