--- src/main/java/com/Horns/Hooves/Employee/exception/EmployeeNotFoundException.java
@@ -1,26 +1,0 @@
-package com.Horns.Hooves.Employee.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class EmployeeNotFoundException extends RuntimeException{
-    public EmployeeNotFoundException() {
-    }
-
-    public EmployeeNotFoundException(String message) {
-        super(message);
-    }
-
-    public EmployeeNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public EmployeeNotFoundException(Throwable cause) {
-        super(cause);
-    }
-
-    public EmployeeNotFoundException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
