--- src/main/java/com/Horns/Hooves/Employee/exception/EmployeeStorageIsFullException.java
@@ -1,22 +1,0 @@
-package com.Horns.Hooves.Employee.exception;
-
-public class EmployeeStorageIsFullException extends RuntimeException {
-    public EmployeeStorageIsFullException() {
-    }
-
-    public EmployeeStorageIsFullException(String message) {
-        super(message);
-    }
-
-    public EmployeeStorageIsFullException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public EmployeeStorageIsFullException(Throwable cause) {
-        super(cause);
-    }
-
-    public EmployeeStorageIsFullException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
