--- src/main/java/com/Horns/Hooves/Employee/service/DepartmentServiceInterf.java
@@ -1,8 +1,0 @@
-package com.Horns.Hooves.Employee.service;
-
-import com.Horns.Hooves.Employee.object.Employee;
-
-public interface DepartmentServiceInterf {
-
-    Employee getMaxSalaryEmployeeByDepartmentNumber(int departmentId);
-}
