--- src/main/java/com/Horns/Hooves/Employee/service/EmployeeServiceInterf.java
@@ -1,13 +1,0 @@
-package com.Horns.Hooves.Employee.service;
-
-public interface EmployeeServiceInterf {
-
-    // Добавление сотрудника
-    String addEmployee(String name, String surname, int salary, int departmentId);
-
-    // Удаление сотрудника
-   String delEmployee(String name, String surname);
-
-    // Поиск сотрудника
-    String findEmployee(String name, String surname);
-}
